+++ RestFlow.API/Controllers/TableController.cs
@@ -17,6 +17,10 @@ namespace RestFlow.API.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<Table>> GetById(int id)
         {
+            if (id == 0)
+            {
+                return BadRequest("Table id is null");
+            }
             var table = await _tableService.GetById(id);
             if (table == null)
             {
@@ -26,22 +30,38 @@ namespace RestFlow.API.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Table>>> GetAll()
+        public async Task<ActionResult<IEnumerable<Table>>> GetAll(int restaurantId)
         {
-            var tables = await _tableService.GetAll();
+            if (restaurantId == 0)
+            {
+                return BadRequest("Restaurant id is null");
+            }
+            var tables = await _tableService.GetAllByRestaurantId(restaurantId);
+            if (tables == null)
+            {
+                return NotFound();
+            }
             return Ok(tables);
         }
 
         [HttpPost]
         public async Task<ActionResult> Create(TableDTO table)
         {
-            await _tableService.Add(table.TableNumber, table.Capacity);
+            if (table == null)
+            {
+                return BadRequest("Table is null");
+            }
+            await _tableService.Add(table.TableNumber, table.Capacity, table.RestaurantId);
             return Ok(table);
         }
 
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(int id)
         {
+            if (id == 0)
+            {
+                return BadRequest("Table id is null");
+            }
             await _tableService.Delete(id);
             return NoContent();
         }
